+++ src/main/java/card/CardHand.java
@@ -1,7 +1,5 @@
-package user;
+package card;
 
-import card.Card;
-import card.CardRank;
 import java.util.ArrayList;
 import java.util.List;
