+++ src/test/java/card/CardDeckTest.java
@@ -1,4 +1,4 @@
-package domain;
+package card;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
