+++ src/test/java/user/DealerTest.java
@@ -1,4 +1,4 @@
-package domain.user;
+package user;
 
 import card.Card;
 import card.CardDeck;
